feat: [#137] 이미지 업로드 구현
