+++ FITple-Frontend/data/SearchMainApi.jsx
@@ -11,9 +11,10 @@ export const searchMain = async (category, cursorId, size) => {
 
     const response = await fetch(url, {
       method: "GET",
-      // headers: {
-      //   "Content-Type": "application/json",
-      // },
+      headers: {
+        "Content-Type": "application/json",
+      },
+      credentials: "include",
     });
 
     if (!response.ok) {
