+++ FITple-Frontend/src/components/RegisterInfo/RegisterInfo.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const RegisterInfo = () => {
+  return <div></div>;
+};
+
+export default RegisterInfo;
